requirements.txt
streamlit
numpy
opencv-python-headless
deepface
tensorflow
tf-keras
Pillow